Day 1 solution
Rust is probably not the go-to language for string parsing and
formatting, but oh well. This wasn't too painful, even with my lack of
rust knowledge.

This contains the solution specifically for the second part of the day
two solution. For the first part, rip out the pseudo-trie and just use
the digit code.