Day 8: Both puzzles
This was a fun one! We could have had more fun with the graph, but this
is probably sufficiently easy.

The solution for puzzle 1 can be easily pulled out of this one, and I
forgot to commit it before embarking on part 2.

It took me a bit to figure out the right way to calculate this was with
the LCM - I had it running through all of them sequentially, and for
some reason that still didn't terminate after 30 minutes :)